improve ergonomics of integer, float, color and boolean formats slightly
I find it quite annoying to keep building the home descriptions and one
of the more annoying things I find is describing properties. This change
improves the situation slightly by making it quite a bit more
straightforward to describe common integer, float and boolean formats.
In particular now instead of having to write out the entire
`HomePropertyFormat::SomeType`, `IntegerRange`, `FloatRange`,
`BooleanFormat` etc can be passed in as an argument directly. For
integers and floats a furher ergonomics are achieved by allowing an even
more common case (range without step) to be provided by using Rust's
built-in ranges.

For color format it might make sense to explore making it a bitfield
rather than a vector of enums. Among other things it would enforce that
the same colour does not appear twice. But even if that's done in the
future these From implementations would still be implementable just
fine, so they seemed fine to add.